Add files via upload
Exercise can lessen your risk of developing significant illnesses like coronary heart disease, stroke, type 2 diabetes,
and cancer as well as your risk of dying young by as much as 30%. This claim makes exercising sound like a great
task to do, but there's another advantage that you're missing. While exercising in a clean environment is
advantageous, doing so in a polluted setting has more of an adverse effect. Our research fundamentally focuses on
educating people about their health, the effects of pollution on them, and the adverse effects of exercising in unhealthy
environments. By outlining where to go and what to do to keep healthy and prevent diseases, it will guide you towards
leading a healthy lifestyle.
The health sector is a significant industry with plenty of space for improvement. The fact that it directly impacts us,
makes it important. If a gap is discovered, it must be filled in as soon as possible. This industry initially drew our
attention while we were looking for a gap, and pollution was the main detractor. After investigating this area, we came
up with our problem statement.